requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0
pyyaml>=6.0

# Jersey OCR (optional — set jersey_ocr.enabled: false in config to skip)
# paddlepaddle>=2.5.0
# paddleocr>=2.7.0
